// src/app/layout/dashboard/employee-information/models/get-employee-role.model.ts
export interface GetEmployeeRole {
  nationalId: number;
  district1: number;
  district2: number;
  district3: number;
  district4: number;
  district1Description: string;
  district2Description: string;
  district3Description: null;
  district4Description: null;
  tableADistrict1Job: number;
  tableADistrict1FromJob: Date;
  tableADistrict1UntilJob: Date;
  tableADistrict2Job: number;
  tableADistrict2FromJob: Date;
  tableADistrict2UntilJob: Date;
  tableADistrict3Job: number;
  tableADistrict3FromJob: Date;
  tableADistrict3UntilJob: Date;
  tableADistrict4Job: number;
  tableADistrict4FromJob: Date;
  tableADistrict4UntilJob: Date;
  tableBDistrict1Job: number;
  tableBDistrict1FromJob: Date;
  tableBDistrict1UntilJob: Date;
  tableBDistrict1Minister: string;
  tableBDistrict2Job: number;
  tableBDistrict2FromJob: Date;
  tableBDistrict2UntilJob: Date;
  tableBDistrict2Minister: string;
  tableBDistrict3Job: number;
  tableBDistrict3FromJob: Date;
  tableBDistrict3UntilJob: Date;
  tableBDistrict3Minister: string;
  tableBDistrict4Job: number;
  tableBDistrict4FromJob: Date;
  tableBDistrict4UntilJob: Date;
  tableBDistrict4Minister: string;
  tableCDistrict1AJob: number;
  tableCDistrict1AFromJob: Date;
  tableCDistrict1AUntilJob: Date;
  tableCDistrict1BJob: number;
  tableCDistrict1BFromJob: Date;
  tableCDistrict1BUntilJob: Date;
  tableCDistrict2AJob: number;
  tableCDistrict2AFromJob: Date;
  tableCDistrict2AUntilJob: Date;
  tableCDistrict2BJob: number;
  tableCDistrict2BFromJob: Date;
  tableCDistrict2BUntilJob: Date;
  tableCDistrict3AJob: number;
  tableCDistrict3AFromJob: Date;
  tableCDistrict3AUntilJob: Date;
  tableCDistrict3BJob: number;
  tableCDistrict3BFromJob: Date;
  tableCDistrict3BUntilJob: Date;
  tableCDistrict4AJob: number;
  tableCDistrict4AFromJob: Date;
  tableCDistrict4AUntilJob: Date;
  tableCDistrict4BJob: number;
  tableCDistrict4BFromJob: Date;
  tableCDistrict4BUntilJob: Date;
  id: number;
}